src/main/java/io/redintro/hexbike/domain/User.java
package io.redintro.hexbike.domain;

import java.util.UUID;

public class User {
    private final UUID id;
    private final String username;
    private final String password;
    private final String role;

    public User(UUID id, String username, String password, String role) {
        this.id = id;
        this.username = username;
        this.password = password;
        this.role = role;
    }

    public UUID getId() {
        return id;
    }

    public String getUsername() {
        return username;
    }

    public String getPassword() {
        return password;
    }

    public String getRole() {
        return role;
    }
}
